Add two new roots
Add "Visa Public RSA Root CA" to all of our logs, as it is trusted
by Microsoft but not yet included in the CCADB export for some reason.

Add "STAGING - Aretiico TLS Root CA G1" to Sapling and Twig, as it is
the staging root of a CA which hopes to apply for root program inclusion
in the future.